+++ Api/Glass.API.Backend/Helper/Produtos/Estrategias/CalculoTotal/IValidacaoCalculoTotal.cs
@@ -35,6 +35,6 @@ namespace Glass.API.Backend.Helper.Produtos.Estrategias.CalculoTotal
         /// </summary>
         /// <param name="dadosAdicionaisValidacao">Os dados adicionais para a validação do produto, se necessário.</param>
         /// <returns>O percentual de desconto por quantidade, se encontrado.</returns>
-        double ObterPercentualDescontoQuantidade(string dadosAdicionaisValidacao);
+        decimal ObterPercentualDescontoQuantidade(string dadosAdicionaisValidacao);
     }
 }
